Switch api-assets to git submodule
Closes #17

+++ frontend/assets.py
@@ -2,66 +2,51 @@
 
 import argparse
 import glob
+import itertools
 import os
-import requests
 import shutil
-import time
+import subprocess
 from PIL import Image
 
-DOWNLOAD_DIR = 'src/api-assets'
-DATA_FILES = [
-    'dictionaries/telemetry/common/mapName.json',
-    'dictionaries/telemetry/item/itemId.json',
-    'dictionaries/telemetry/vehicle/vehicleId.json',
-    'dictionaries/telemetry/damageCauserName.json',
-]
+ASSETS_DIR = 'src/api-assets'
 
-MAP_DIR = 'public/assets/maps'
-TILE_PATH_FORMAT = 'public/assets/maps/{map_key}/{x}_{y}.jpg'
+# Maps
+MAP_SRC_DIR = os.path.join(ASSETS_DIR, 'Assets/Maps')
+TILE_MAP_DEST_DIR = 'public/assets/maps'
 TILE_SIZE = 512
 ZOOM_LEVELS = 5
 
-
-def download(file, dest_dir=DOWNLOAD_DIR):
-    local_path = os.path.join(dest_dir, file)
-
-    headers = {}
-    # Attempt to check the local modified time to prevent downloading identical files
-    try:
-        # Only download the file if the remote copy is newer. It seems like this doesn't actually
-        # work because GitHub doens't respect the If-Modified-Since header
-        mod_time = os.path.getmtime(local_path)
-        timestamp = time.strftime('%a, %d %b %Y %H:%M:%S GMT', time.gmtime(mod_time))
-        headers['If-Modified-Since'] = timestamp
-    except FileNotFoundError:
-        pass
-
-    print(f"Downloading '{file}'")
-    r = requests.get(args.repo + file, headers=headers)
-    r.raise_for_status()
-    if r.status_code == 304:
-        print("  Skipping, already latest")
-    else:
-        print(f"  Saving to '{local_path}'")
-        os.makedirs(os.path.dirname(local_path), mode=0o755, exist_ok=True)
-        with open(local_path, 'wb') as f:
-            f.write(r.content)
-    return local_path
+DELETE_PROMPT = "{path} already exists, do you want to delete it?"
 
 
-def download_all():
-    os.chdir(os.path.dirname(os.path.realpath(__file__)))  # Move to this script's path
+def confirm_prompt(prompt, padding=0, default_yes=True):
+    if args.yes:
+        return True
+    padding_str = ' ' * padding
+    response = input(f"{padding_str}{prompt} [{'Y/n' if default_yes else 'y/N'}] ")
+    return response.startswith('y') or (not response.startswith('n') and default_yes)
 
-    # Download data files
-    for file in DATA_FILES:
-        download(file)
+
+def pull_latest(**kwargs):
+    os.chdir(ASSETS_DIR)
+    old_commit = subprocess.check_output(['git', 'rev-parse', 'HEAD']).strip()
+    subprocess.run(['git', 'submodule', 'update', '--recursive', '--remote', '--', '.'])
+    diff = subprocess.check_output(['git', 'diff', '--name-only', old_commit]).decode()
+    print(diff or "No Changes")
 
 
 def tile_for_all_zooms(image_path, output_dir):
-    print(f"Tiling '{image_path}', saving to '{output_dir}'")
+    print(f"Tiling {image_path} => {output_dir}")
+
+    if os.path.exists(output_dir) and not confirm_prompt(DELETE_PROMPT.format(path=output_dir),
+                                                         padding=2):
+        print(f"  Skipping {image_path}")
+        return
+
+    # Delete the directory (if present) and re-make it
     try:
         shutil.rmtree(output_dir)
-        print(f"Deleted existing '{output_dir}' (RIP)")
+        print(f"  Deleted {output_dir}")
     except FileNotFoundError:
         pass
     os.makedirs(output_dir)
@@ -84,32 +69,45 @@ def tile_image(image, output_dir):
     width, height = image.size
     for x in range(0, width, TILE_SIZE):
         for y in range(0, height, TILE_SIZE):
-            box = (x, y, x + TILE_SIZE, y + TILE_SIZE)
-            tile_path = os.path.join(output_dir, f'{x // TILE_SIZE}_{y // TILE_SIZE}.jpg')
+            box = (x, y, x + TILE_SIZE, y + TILE_SIZE)  # The portion of the parent image to crop
+
+            # Build a path for the new file
+            tile_x = x // TILE_SIZE
+            tile_y = y // TILE_SIZE
+            tile_path = os.path.join(output_dir, f'{tile_x}_{tile_y}.jpg')
+
             cropped = image.crop(box)
             cropped.save(tile_path, quality=95, subsampling=0)  # No compression
 
 
-def tile_all():
-    glob_path = os.path.join(args.map_dir, '*.jpg')
-    for map_file in glob.glob(glob_path):
+def tile_all(output_dir, **kwargs):
+    exts = ('*.jpg', '*.png')
+    all_files = itertools.chain.from_iterable(glob.glob(os.path.join(MAP_SRC_DIR, ext))
+                                              for ext in exts)
+    files = (f for f in all_files if 'lowres' not in f)  # Filter out lowres images
+    for map_file in files:
         map_key = os.path.splitext(os.path.basename(map_file))[0]
-        tile_for_all_zooms(map_file, os.path.join(MAP_DIR, map_key))
-
-
-ACTIONS = {
-    'download': download_all,
-    'tile': tile_all,
-}
+        tile_for_all_zooms(map_file, os.path.join(output_dir, map_key))
 
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser("Script to download and save PUBG assets")
-    parser.add_argument('action', choices=ACTIONS.keys())
-    parser.add_argument('--repo', '-r', default='https://github.com/pubg/api-assets/raw/master/',
-                        help="URL of the repo to download from")
-    parser.add_argument('--map-dir', '-m', default=MAP_DIR,
-                        help="Directory containing map images to tile")
+    parser.add_argument('--yes', '-y', action='store_true',
+                        help="Answer yes to all confirmation prompts")
+    subparsers = parser.add_subparsers()
+
+    pull_parser = subparsers.add_parser('pull', help="Pull the latest version of the assets repo")
+    pull_parser.set_defaults(func=pull_latest)
+
+    tile_parser = subparsers.add_parser('tile', help="Break map images into tiles")
+    tile_parser.add_argument('--output-dir', '-o', default=TILE_MAP_DEST_DIR,
+                             help="Directory to output tiled maps to")
+    tile_parser.set_defaults(func=tile_all)
+
+    script_dir = os.path.dirname(os.path.realpath(__file__))
+    os.chdir(script_dir)
+
     args = parser.parse_args()
-    func = ACTIONS[args.action]
-    func()
+    argd = vars(args)
+    func = argd.pop('func')
+    func(**argd)  # Run the function for the action
